Added ability to select files, folders to add to the list of files. Cleanup code and refactored methods to expose static functions instead of code in the event

+++ Source/RFUpdater.Models/ModFile.cs
@@ -8,7 +8,7 @@ namespace RFUpdater.Models
         public string FileName { get; set; }
         public string FilePath { get; set; }
         public UInt32 FileHash { get; set; }
-        public UInt32 FileSize { get; set; }
+        public long FileSize { get; set; }
         public Boolean Protected { get; set; }
 
         public string GetFullPath()
